fix: relocate skipFuseValidationError to avoid tanking coverage
